rename `onboard` -> `tunnel`.
It now does more than onboarding, but a load of SSH stuff, so it needs a broader name.
