+++ internal/tunnel/onboard.go
@@ -1,4 +1,4 @@
-package onboard
+package tunnel
 
 import (
 	"bytes"
